python/python_aws/requirements.txt
pendulum
boto3
python-dotenv